Moving scm-input handling to separate module, more attempts at increased test speed
